% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R
\name{make_mgcvRE}
\alias{make_mgcvRE}
\title{Create "mgcv style" random effect terms (internal)}
\usage{
make_mgcvRE(
  term,
  subj_id,
  order_id,
  wiggle_subj = TRUE,
  set_k_by = "mgcv",
  bs = "tp",
  k_threshold = 0.25,
  data = NULL
)
}
\arguments{
\item{term}{the \code{s()} term that needs to be processed.}
}
\value{
A new \code{s()} term with additional parameter
}
\description{
This function adds arguments to \code{s()} term that is use in \code{{mgcv}} formulas.
Better leave this function internal.
}
\details{
The "smooth term" \code{s()} that enters this function should only contain factors, e.g., \code{s(id)}, \code{s(item)}.
See \code{\link[=wi_mgcv]{wi_mgcv()}} for details of other arguments.
}
\examples{
## create a 'clean' by-subject random effect term
bysubj_int <- quote(s(id))

## standard by-subject random effect
make_mgcvRE(term = bysubj_int,
            subj_id = quote(id),
            item_id = quote(item_id),
            wiggle_subj = FALSE)

## wiggle by-subject effect
make_mgcvRE(term = bysubj_int,
            subj_id = quote(id),
            item_id = quote(item_id),
            wiggle_subj = TRUE, set_k_by = "mgcv",
            bs = "tp", k_threshold = 0.25)

## by-subject random slope
make_mgcvRE(term = quote(s(id, A)),
            subj_id = quote(id),
            item_id = quote(item_id))

}
\seealso{
\code{\link[=wi_mgcv]{wi_mgcv()}}
}
\keyword{internal}
