% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{possible_subblocks}
\alias{possible_subblocks}
\title{Set of possible sub-blocks}
\usage{
possible_subblocks(n_reps)
}
\arguments{
\item{n_reps}{Number of repetitions of each factor level.}
}
\value{
A vector with the possible number of sub-blocks.
}
\description{
Calculates the set of possible sub-blocks for given values of
\code{n_reps} (number of repetitions of each factor level).
}
\details{
This is a basic function that wraps the function \code{\link[numbers:divisors]{numbers::divisors()}}.
}
\examples{
possible_subblocks(48)

}
