% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomization.R
\name{psr}
\alias{psr}
\title{Permuted-subblock randomization (PSR) for one-factor designs}
\usage{
psr(levels, n_subblocks, n_reps)
}
\arguments{
\item{levels}{Set of condition levels (labels).}

\item{n_subblocks}{Number of desired subblocks.}

\item{n_reps}{Number of repetitions of each level.}
}
\value{
A character vector of length \code{n_reps} times \code{length(levels)}.
}
\description{
Given a vector of condition labels, \code{psr()} outputs a
permuted-subblock randomization for a single participant.
}
\examples{
psr(c("A", "B", "C"), 6, 12)

## do the same but for four participants
lapply(seq_len(4),
       \(.x) psr(c("A", "B", "C"), 6, 12))

}
\seealso{
\code{\link[=psr_stimuli]{psr_stimuli()}} for applying PSR to a table with stimuli.
}
