% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomization.R
\name{psr_2x2_stimuli}
\alias{psr_2x2_stimuli}
\title{Permuted-subblock randomization (PSR) for 2x2 factorial designs with stimuli}
\usage{
psr_2x2_stimuli(
  stim_table,
  IVs,
  algorithm = c("max", "circular", "eight"),
  n_part = 1L,
  sb_varname = "sb_no",
  part_varname = "PID"
)
}
\arguments{
\item{stim_table}{Table containing the stimuli for the experiment.}

\item{IVs}{Vector with quoted names of the variables in \code{stim_table}
corresponding to the independent variables in the study.}

\item{algorithm}{The algorithm to order conditions, \code{max} uses PSR-max, \code{circular}
uses PSR-C, and "eight" uses PSR-E. Use \code{con_2x2()} to get an intuitive
example of how orders are arranged by PSR-C and PSR-E.}

\item{n_part}{Number of desired participants.}

\item{sb_varname}{Name to give the subblocks variable in the output.}

\item{part_varname}{Name to give the participant variable in the output.}
}
\value{
A data frame with \code{nrow(stim_table) * n_part} rows representing
permuted-subblock randomization applied to \code{stim_table}. Output includes
variables \code{PID} identifying individual participants and \code{sb_no}
identifying subblocks within participants.
}
\description{
Given a table with a list of the stimuli presented to each participant,
outputs a table with stimuli organized into subblocks.
}
\details{
When \code{algorithm = c("c", "e")} and \code{np} is not a multiple of 4, sequences
would still be randomized, but a warning message would be generated reminding
this is not a fully counterbalanced design.
}
\examples{

psr_stimuli(stim_table = stroop_stimuli_factorial,
            IVs = c("font_color", "congruency"),
            algorithm = "circular",
            n_part = 4L,
            sb_varname = "sb_no",
            part_varname = "PID")

}
