% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomization.R
\name{psr_stimuli}
\alias{psr_stimuli}
\title{Permuted-subblock randomization (PSR) with stimuli}
\usage{
psr_stimuli(
  stim_table,
  IVs,
  n_subblocks,
  n_part = 1L,
  sb_varname = "sb_no",
  part_varname = "PID"
)
}
\arguments{
\item{stim_table}{Table containing the stimuli for the experiment.}

\item{IVs}{Vector with quoted names of the variables in \code{stim_table}
corresponding to the independent variables in the study.}

\item{n_subblocks}{Number of desired subblocks.}

\item{n_part}{Number of desired participants.}

\item{sb_varname}{Name to give the subblocks variable in the output.}

\item{part_varname}{Name to give the participant variable in the output.}
}
\value{
A data frame with \code{nrow(stim_table) * n_part} rows representing
permuted-subblock randomization applied to \code{stim_table}. Output includes
variables \code{PID} identifying individual participants and \code{sb_no}
identifying subblocks within participants.
}
\description{
Given a table with a list of the stimuli presented to each participant,
outputs a table with stimuli organized into subblocks.
}
\examples{

psr_stimuli(stroop_stimuli, "congruency", 12)

psr_stimuli(stroop_stimuli, c("font_color", "congruency"), 3)

}
