% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{streak_table}
\alias{streak_table}
\title{Tabulate streaks in a vector}
\usage{
streak_table(data, varname, pid_name = "PID")
}
\arguments{
\item{data}{A data frame.}

\item{varname}{The name of the variable you want to check for streaks.}

\item{pid_name}{The name of the variable identifying individual participants.}
}
\value{
A data frame representing a contingency table with two variables,
\code{streak_length}, the length of the streak, and \code{n}, the number of times it
occurs for the variable \code{varname}.
}
\description{
Given a vector \code{x}, produce a frequency table of the number of
streaks (i.e., unbroken sequences of the same value).
}
\examples{

stroop1 <- psr_stimuli(stroop_stimuli, "congruency", 1, 16)
streak_table(stroop1, "congruency", "PID")

stroop12 <- psr_stimuli(stroop_stimuli, "congruency", 12, 16)
streak_table(stroop12, "congruency", "PID")

}
