% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stroop_stimuli}
\alias{stroop_stimuli}
\title{Sample stimuli for a Stroop experiment}
\format{
\subsection{\code{stroop_stimuli}}{

A data frame with 48 rows and 4 columns.
\describe{
\item{stimulus_id}{A factor uniquely identifying each stimulus in the dataset.}
\item{word}{Identity of the printed word (blue, green, red, or yellow).}
\item{font_color}{Font color of the word (blue, green, red, or yellow).}
\item{congruency}{Whether the word and font color are congruent or incongruent.}
}
}
}
\usage{
stroop_stimuli
}
\description{
Description of a set of stimuli that could be used in a
Stroop-style cognitive experiment.
}
\keyword{datasets}
