% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stroop_stimuli_factorial}
\alias{stroop_stimuli_factorial}
\title{Sample stimuli for a factorial Stroop experiment}
\format{
\subsection{\code{stroop_stimuli_factorial}}{

A data frame with 48 rows and 5 columns.
\describe{
\item{stimulus_id}{A factor uniquely identifying each stimulus in the dataset.}
\item{word}{Identity of the printed word (blue, green, red, or brown).}
\item{font_color}{Font color of the word (blue, green, red, or brown).}
\item{congruency}{Whether the word and font color are congruent or incongruent.}
\item{positions}{Whether this stimuli requires participants to stand or sit.}
}
}
}
\usage{
stroop_stimuli_factorial
}
\description{
Description of a set of stimuli that could be used in a
Stroop-style cognitive experiment.
}
\keyword{datasets}
