% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomization.R
\name{walk_2x2}
\alias{walk_2x2}
\title{Arrange stimuli for a 2x2 factorial design, in circular or "eight" patterns.}
\usage{
walk_2x2(np = 4, nr = 4, facR = c("A", "a"), facC = c("B", "b"), method = "e")
}
\arguments{
\item{np}{Number of desired participants.}

\item{nr}{Number of repetitions of each level.}

\item{facR}{A vector containing two labels of the first (row) factor.}

\item{facC}{A vector containing two labels of the second (column) factor.}

\item{method}{The algorithm to order conditions, \code{c} represents "circular", \code{e} represents
"figure-8" pattern. Use \code{con_2x2()} to get an intuitive example.}
}
\value{
A list with the length of \code{np} with each element containing a sequence with
the length of \code{nr}.
}
\description{
Given two factors and their levels, \code{walk_2x2()} outputs a sequences of combined
conditions and arrange their order in circular or "figure-eight" patterns.
}
\details{
When multiple participants are desired, \code{walk_2x2()} counterbalances sequences
across participants.

Original condition order can be inspected with \code{con_2x2()} function.
}
\examples{
walk_2x2(np = 1, nr = 4,
         facR = c("A", "a"), facC = c("B", "b"),
         method = "c")

}
\seealso{
\code{\link[=psr_stimuli]{psr_stimuli()}} for applying PSR to a table with stimuli in
one-factordesign.

\code{\link[=psr_2x2_stimuli]{psr_2x2_stimuli()}} for applying PSR to a table with stimuli in
2x2 factorial design.
}
