% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R
\name{wi_mgcv}
\alias{wi_mgcv}
\title{Translate \code{{lme4}} formula to \code{{mgcv}} style and specify a wiggly by-subject random intercept}
\usage{
wi_mgcv(
  formula,
  subj_id,
  order_id,
  wiggle_subj = TRUE,
  set_k_by = "explan",
  bs = "tp",
  k_threshold = 0.25,
  data = NULL
)
}
\arguments{
\item{formula}{a \code{{lme4}}-style formula}

\item{subj_id}{the column in data (or its name) to identify unique subjects}

\item{order_id}{the column in data (or its name) to identify trial orders. Required only when \code{wiggle_subj = TRUE}.}

\item{wiggle_subj}{whether 'wiggly by_subject random intercept' should be included; default is \code{TRUE}}

\item{set_k_by}{how the number of basis function \code{k} is determined. Can be either \code{"mgcv"} or \code{"explan"}. See details for further description.}

\item{bs}{'basis smooth' used to specify 'wigglyness' of factor smooth. This argument follows \code{\link[mgcv:s]{mgcv::s()}} settings and works only when \code{wiggle_subj = TRUE}; the default value is \code{"tp"}}

\item{k_threshold}{A value to determine the number of basis functions. This argument works only when \code{wiggle_subj = TRUE}, \code{set_k_by = "explan"}, and \code{data} is specified; the default value is 0.25. See details.}

\item{data}{a formula-relevant data frame that contains corresponding variable names in \code{formula}.}
}
\value{
a new formula that is appropriate in \code{\link[mgcv:bam]{mgcv::bam()}} or \code{\link[mgcv:gam]{mgcv::gam()}} with expanded random effect specification.
}
\description{
This function translates a \code{{lme4}} mixed-effect formula (whose random effects are defined by \code{|} notations)
and translate it to a {mgcv} mixed-effect formula (whose random effects are defined by \code{s()}).
}
\details{
\code{\link[=wi_mgcv]{wi_mgcv()}} is mainly used to efficiently convert a {lme4} mixed-effect formula (which usually does not support wiggly by-subject random intercept) to a {mgcv} mixed-effect formula (does). By specifying subject (participant) id column and order id column, \code{wi_mgcv()} identifies static by-subject random intercept and convert it to 'factor smooths'. By doing so, by-subject random intercept can wiggle over time (specified by \code{order_id}).

Note that, during conversion, any correlated random effect structure (e.g., correlated random slopes and random intercepts such as \code{(1 + dv | id)}) would be forced to uncorrelated to meet the requirements in \code{{mgcv}} context. It is recommended to use double-bar notations to specify uncorrelated random effects in your \code{{lme4}} formula, such as \code{(dv1 * dv2 || id)}.

Only when \code{wiggle_subj = TRUE}, arguments \code{set_k_by = "explan"} and \code{k_threshold} become meaningful. When \code{wiggle_subj = FALSE}, the function simply produces a formula that fits Linear Mixed-Effect Model in \verb{\{mgcv\}.}

If \code{set_k_by = "mgcv"}, the number of basis functions \code{k} is determined by package \code{{mgcv}} default settings; if \code{set_k_by = "explan"}, argument \code{data} is required, as \code{"explan"} uses data information (maximal trial number) to determine number of basis function. The default is 25\% of trials of each subject.
}
\examples{

## 1: one-factor design
n_part <- 4
dat_1f <- psr_stimuli(stroop_stimuli, "congruency",
                      n_subblocks = 1, n_part = n_part) |>
    cbind(order = rep(1:nrow(stroop_stimuli),
                      times = n_part),
          dv = rnorm(n = nrow(stroop_stimuli) * n_part))

## both lme4::lmer() and mgcv::bam() are coding-sensitive
dat_1f[["PID"]] <- factor(dat_1f[["PID"]])
dat_1f[["congruency"]] <- ifelse(dat_1f[["congruency"]] == "congruent",
                                 yes = -1/2, no = 1/2)
mod_1f <- dv ~ congruency +
    (congruency || PID)

## default settings: GAMMs with 'wiggly intercept' (factor smooth)
mod_mgcv <- wi_mgcv(formula = mod_1f, subj_id = PID,
                    order_id = order, data = dat_1f)
mgcv::bam(mod_mgcv, data = dat_1f) |> summary()

## equivalent to the original lme4 model, but available for `mgcv::gam()`
mod_mgcv_e <- wi_mgcv(formula = mod_1f, subj_id = PID,
                      wiggle_subj = FALSE)
lme4::lmer(mod_1f, data = dat_1f) |> summary()
mgcv::bam(mod_mgcv_e, data = dat_1f) |> summary()

## `data` is ignored when using default mgcv sttings to determine k...
wi_mgcv(formula = mod_1f, subj_id = PID,
        order_id = order,
        set_k_by = "mgcv")
## ... but not when using `explan` method
wi_mgcv(formula = mod, subj_id = PID,
        order_id = order,
        set_k_by = "explan",
        data = dat_1f)

## 2: two-factor design, both within-subject
dat_2f_ww <- psr_2x2_stimuli(
    stim_table = stroop_stimuli_factorial,
    IVs= c("congruency", "positions"),
    algorithm = "circular",
    n_part = n_part) |>
    cbind(order = rep(1:nrow(stroop_stimuli_factorial),
                      times = n_part),
          dv = rnorm(n = nrow(stroop_stimuli_factorial) * n_part))
dat_2f_ww[["PID"]] <- factor(dat_2f_ww[["PID"]])
dat_2f_ww[["congruency"]] <- ifelse(dat_2f_ww[["congruency"]] == "congruent",
                                    yes = -1/2, no = 1/2)
dat_2f_ww[["positions"]] <- ifelse(dat_2f_ww[["positions"]] == "standing",
                                   yes = -1/2, no = 1/2)

mod_2f_ww <- dv ~ congruency * positions +
    (congruency * positions || PID)
mod_mgcv_2f_ww <- wi_mgcv(formula = mod_2f_ww, subj_id = PID,
                       order_id = order, data = dat_2f_ww)
mgcv::bam(mod_mgcv_2f_ww, data = dat_2f_ww) |> summary()


mod_mgcv_2f_ww_e <- wi_mgcv(formula = mod_2f_ww, subj_id = PID,
                            order_id = order, wiggle_subj = FALSE)

## equivalent
lme4::lmer(mod_2f_ww, data = dat_2f_ww) |> summary()
mgcv::bam(mod_mgcv_2f_ww_e, data = dat_2f_ww) |> summary()

## 3: two-factor design, one within-subject and one between-subject
dat_2f_wb <- psr_stimuli(stroop_stimuli, "congruency",
                         n_subblocks = 1, n_part = n_part) |>
    cbind(group = factor(rep(1:2, each = nrow(stroop_stimuli) * 2)),
          order = rep(1:nrow(stroop_stimuli),
                      times = n_part),
          dv = rnorm(n = nrow(stroop_stimuli) * n_part))
dat_2f_wb[["PID"]] <- factor(dat_2f_wb[["PID"]])
dat_2f_wb[["group"]] <- factor(dat_2f_wb[["group"]])
dat_2f_wb[["congruency"]] <- ifelse(dat_2f_wb[["congruency"]] == "congruent",
                                    yes = -1/2, no = 1/2)

mod_2f_wb <- dv ~ congruency * group +
    (congruency * group || PID) +
    (group || stimulus_id)

mod_mgcv_2f_wb <- wi_mgcv(formula = mod_2f_wb, subj_id = PID,
                          order_id = order, data = dat_2f_wb)
mgcv::gam(mod_mgcv_2f_wb, data = dat_2f_wb) |> summary()

mod_mgcv_2f_wb_e <- wi_mgcv(formula = mod_2f_wb,
                            subj_id = PID, wiggle_subj = FALSE,
                            order_id = order,
                            data = dat_2f_wb)

## use gam() when experiencing "non-conformable arrays" error as it is a known bug
lme4::lmer(mod_2f_wb, data = dat_2f_wb) |> summary()
mgcv::bam(mod_mgcv_2f_wb_e, data = dat_2f_wb, method = "REML") |> summary()

## 4. formulae
## these formulae are acceptable...
dv ~ A + (1 | id) ## random-intercept-only model
dv ~ A + (1 | id) + (0 + A | id)
dv ~ A + (A || id)
dv ~ A * B + (A * B || id)
dv ~ A * B + (A * B || id) + (B || order)

## ...while these produce warnings
dv ~ A + (A | id) ## correlated ran int/slp
dv ~ A + (0 + A | id) ## ran-slp only model
dv ~ A + (-1 + A | id) ## ran-slp only model, alternative
dv ~ A * B + (A * B | id) + (B | order)
dv ~ A * B + (A * B | id) + (B || order)
dv ~ A * B + (A * B || id) + (B | order)
dv ~ A * B + (0 + A * B | id) + (B | order)
}
