% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{con_2x2}
\alias{con_2x2}
\title{Construct a 2x2 table for conditions.
Calculates the set of possible sub-blocks for given values of
\code{n_reps} (number of repetitions of each factor level).}
\usage{
con_2x2(facR = c("A", "a"), facC = c("B", "b"))
}
\arguments{
\item{facR}{A vector containing two labels of the first (row) factor.}

\item{facC}{A vector containing two labels of the second (column) factor.}
}
\value{
A matrix with each entry being one of the conditions, followed by
corresponding numbers to identify relitive positions in PSR-C and PSR-E.
}
\description{
Construct a 2x2 table for conditions.
Calculates the set of possible sub-blocks for given values of
\code{n_reps} (number of repetitions of each factor level).
}
\examples{
con_2x2(facR = c("A", "a"),
        facC = c("B", "b")
}
\seealso{
\code{\link[=walk_2x2]{walk_2x2()}} to order stimuli based on relitive positions.
\code{\link[=psr_2x2_stimuli]{psr_2x2_stimuli()}} to randomize 2x2 factorial design stimuli with PSR.
}
